Add docs. Improve images js organization. Move activestorage key path usage to config.
