+++ app/helpers/application_helper.rb
@@ -6,8 +6,9 @@ module ApplicationHelper
   #
   # @param [String] s3_key A key pointing to an object in S3 (not including the bucket name prefix)
   # @note Keys should generally not have a leading '/' in the key.
+  # @note Constructed manually instead of through ActiveStorage because this image is not tied to a real model
   # @return [String]
   def recipe_bucket_asset_url(s3_key)
-    "https://#{LAZY_KITCHEN_BUCKET}.s3.amazonaws.com/#{s3_key}"
+    "https://#{Rails.configuration.x.application_bucket}.s3.amazonaws.com/#{s3_key}"
   end
 end
