+++ app/models/concerns/active_storage_path.rb
