+++ config/environments/production.rb
@@ -45,6 +45,8 @@ Rails.application.configure do
 
   # Store uploaded files on the local file system (see config/storage.yml for options).
   config.active_storage.service = :amazon
+  # Path prefix for S3 resources item files (such as images) (defaults to dev path)
+  config.x.resource_prefix = 'resources'
 
   # Mount Action Cable outside main process or domain.
   # config.action_cable.mount_path = nil
