--- config/initializers/s3.rb
@@ -1,1 +1,0 @@
-LAZY_KITCHEN_BUCKET = 'lazy-kitchen'
